Refactored inet tests into a standalone test library and test drivers.
